+++ src/SDKs/DataFactory/DataFactory.Tests/JsonSamples/DatasetJsonSamples.cs
@@ -468,5 +468,26 @@ namespace DataFactory.Tests.JsonSamples
         }
     }
 }";
+
+        [JsonSample]
+        public const string SapCloudForCustomerResourceDataset = @"
+{
+    name: ""SapCloudForCustomerResourceDataset"",
+    properties:
+    {
+        type: ""SapCloudForCustomerResource"",
+        linkedServiceName: 
+        {  
+            referenceName : ""ls"",
+            type : ""LinkedServiceReference""
+        },
+        typeProperties:
+        {            
+            path: ""LeadCollection""
+        }
+    }
+}
+";
+
     }
 }